simplify routing setup for single NIC systems
In case there is only a single NIC, skip routing policies. Routing
policies are not needed in that case and the setup created by cloud-netconfig
is known to interfere with overlay network setup used by kubernetes (see also
bsc#1135592). This does not solve the issue entirely but at least restricts it
to multi-NIC systems.