also copy IPv6 routes into interface specific tables
Copy IPv6 (not just IPv4) routes from the default table into interface
specific ones to make sure any specific routing is preserved.
Also make sure that in case there are multiple routes for the same destination
but via different interfaces, the one matching the interface the routing
table is associated with is selected.