improve default gateway guessing
In case there is no gateway given by the DHCP4 server, assume gateway is the
first host of the network (and not simply .1, which may not work for some
netmasks).  In case the netmask is 32 without a gateway specified (current
setup in GCE), parse the additional routes specified by the DHCPv4 server for
the gateway address.